Refactor journey handling by introducing ensurePackJourneyNormalized to ensure pack.journey exists and is properly structured. Update journey editor to utilize this new function for managing journey stops. Remove deprecated journey-editor.js and update related modules for improved journey management.
