+++ src/index.html
@@ -472,7 +472,7 @@
             <option value="heightmap">Heightmap</option>
             <option value="physical">Physical map</option>
             <option value="poi">Places of interest</option>
-            <option value="campaign">Campaign</option>
+            <option value="journeyPath">Journey path</option>
             <option value="military">Military map</option>
             <option value="emblems">Emblems</option>
             <option value="landmass">Pure landmass</option>
@@ -8699,14 +8699,14 @@
     <script defer src="libs/alea.min.js?v1.105.0"></script>
     <script defer src="libs/polylabel.min.js?v1.105.0"></script>
     <script defer src="libs/simplify.js?v1.105.6"></script>
-    <script defer src="modules/ui/layers.js?v=1.122.7"></script>
+    <script defer src="modules/ui/layers.js?v=1.122.8"></script>
     <script defer src="modules/ui/measurers.js?v=1.99.00"></script>
     <script defer src="modules/ui/style-presets.js?v=1.100.01"></script>
     <script defer src="modules/ui/general.js?v=1.120.6"></script>
     <script defer src="modules/ui/options.js?v=1.120.5"></script>
     <script defer src="main.js?v=1.122.2"></script>
 
-    <script defer src="modules/ui/style.js?v=1.120.7"></script>
+    <script defer src="modules/ui/style.js?v=1.120.8"></script>
     <script defer src="modules/ui/editors.js?v=1.120.5"></script>
     <script defer src="modules/ui/tools.js?v=1.122.0"></script>
     <script defer src="modules/ui/world-configurator.js?v=1.120.5"></script>
@@ -8718,7 +8718,6 @@
     <script defer src="modules/ui/temperature-graph.js?v=1.106.6"></script>
     <script defer src="modules/ui/routes-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/routes-creator.js?v=1.120.5"></script>
-    <script defer src="modules/ui/journey-editor.js?v=1.123.3"></script>
     <script defer src="modules/ui/route-group-editor.js?v=1.120.5"></script>
     <script defer src="modules/ui/ice-editor.js?v=1.111.0"></script>
     <script defer src="modules/ui/lakes-editor.js?v=1.120.5"></script>
@@ -8751,7 +8750,7 @@
     <script defer src="libs/rgbquant.min.js"></script>
     <script defer src="libs/jquery.ui.touch-punch.min.js"></script>
     <script defer src="modules/io/save.js?v=1.122.0"></script>
-    <script defer src="modules/io/load.js?v=1.122.2"></script>
+    <script defer src="modules/io/load.js?v=1.122.3"></script>
     <script defer src="modules/io/cloud.js?v=1.114.0"></script>
     <script defer src="modules/io/export.js?v=1.120.5"></script>
   </body>
