+++ src/modules/index.ts
@@ -10,6 +10,7 @@ import "./biomes";
 import "./cultures-generator";
 import "./routes-generator";
 import "./journey-draw";
+import "./journey-editor";
 import "./states-generator";
 import "./zones-generator";
 import "./religions-generator";
